refactor: move the key ring down into Abblix.Jwt
The key ring lived inside Abblix.Oidc.Server, but nothing in it is about being
an OpenID Provider. It mints, seals, stores, rotates and orders JsonWebKeys -
which is the subject of Abblix.Jwt, where its own central contract
IKeyCustodian already lived. Contract and machinery were split across two
packages; this puts the package boundary where the responsibility boundary
already was.

The split falls along a seam the code had already drawn. Nothing that moved
mentions the server's key-provider contract; only the three providers that
CONSUME the ring did, and those stay. So the dependency always pointed the
right way.

What moved to Abblix.Jwt/ExternalKeys: KeyRing, KeyRingRefreshService,
IKeyRingStore, KeyEnvelope, StoredKey, MintedKeys, ProduceFirstOrdering,
CustodianHeldKeys, CustodianTierValidation and the four builders, joined by the
external-key types that were already in Abblix.Jwt but scattered across its
root, Signing and Encryption folders - IKeyCustodian, ExternalKeySigner,
ExternalKeyDecryptor.

What stays in the server: the four implementations of IAuthServiceKeysProvider
and the JWKS packaging. The server asks its provider for keys; the concrete
provider takes them either from OidcOptions or from the ring. Which one answers
is decided by what the registration puts in the container, and that is
unchanged.

KeyRing is now behind IKeyRing, so the wiring around it stays internal and a
consumer depends on a contract rather than a class. The ring hands its keys to
whoever asks; what they are then used for is the asker's business.

The single setting the ring read from OidcOptions moves to KeyRingOptions
beside it. The host-facing API is untouched: AddCustodian().UseKeysInProcess()
reads and behaves exactly as before, because what moved is the package boundary
underneath it, not the surface.

Done now rather than after 2.4 because none of these types have shipped - v2.3
does not contain them - so moving public types between assemblies breaks no
published consumer today. That window closes at the next release.

Solution builds clean, 3491 tests pass.
