+++ src/Abblix.Jwt/Signing/IDataSigner.cs
@@ -27,7 +27,7 @@ namespace Abblix.Jwt.Signing;
 /// keys it owns. Backends compose as peers behind <see cref="CompositeSigner"/>, which asks each in turn
 /// whether it owns the key (<see cref="CanSign"/>) and routes to the first that does: the in-process
 /// <see cref="LocalKeySigner"/> owns keys that carry private material, an external custodian backend
-/// (<see cref="ExternalKeySigner"/>) owns the public-only keys whose <c>kid</c> is its handle. This is the
+/// (<see cref="ExternalKeys.ExternalKeySigner"/>) owns the public-only keys whose <c>kid</c> is its handle. This is the
 /// byte-level counterpart of the token-level <see cref="IJsonWebTokenSigner"/>: it works with bytes, not a
 /// whole token, so an HSM/KMS/vault integration is one more backend and never touches JWS framing.
 /// </summary>
