+++ src/Abblix.Oidc.Server.Azure/KeyVaultClient.cs
@@ -24,6 +24,7 @@ using System.Collections.Concurrent;
 using System.Net;
 using System.Runtime.CompilerServices;
 using Abblix.Jwt;
+using Abblix.Jwt.ExternalKeys;
 using Azure;
 using Azure.Core;
 using Azure.Core.Pipeline;
