+++ src/Abblix.Oidc.Server.Vault/KeyValueStore.cs
@@ -23,7 +23,7 @@
 using System.Globalization;
 using System.Net;
 using System.Text.Json;
-using Abblix.Oidc.Server.Features.ExternalKeys;
+using Abblix.Jwt.ExternalKeys;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
