+++ src/Abblix.Oidc.Server.Vault/TransitCustodian.cs
@@ -25,6 +25,7 @@ using System.Net;
 using System.Runtime.CompilerServices;
 using System.Security.Cryptography;
 using Abblix.Jwt;
+using Abblix.Jwt.ExternalKeys;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
