+++ tests/Abblix.Jwt.UnitTests/Abblix.Jwt.UnitTests.csproj
@@ -20,7 +20,8 @@
         <PackageReference Include="Microsoft.Extensions.Logging" />
         <PackageReference Include="Microsoft.Extensions.Logging.Console" />
         <PackageReference Include="System.IdentityModel.Tokens.Jwt" />
-        <PackageReference Include="xunit.v3.mtp-v2" />
+        <PackageReference Include="Moq" />
+    <PackageReference Include="xunit.v3.mtp-v2" />
     <PackageReference Include="Microsoft.Testing.Extensions.CodeCoverage" />
         <PackageReference Include="Microsoft.Testing.Extensions.TrxReport" />
         <PackageReference Update="SonarAnalyzer.CSharp">
