+++ tests/Abblix.Jwt.UnitTests/ExternalSignerTests.cs
@@ -22,6 +22,7 @@
 
 using System.Runtime.CompilerServices;
 using System.Security.Cryptography;
+using Abblix.Jwt.ExternalKeys;
 using Abblix.Jwt.Signing;
 using Microsoft.Extensions.DependencyInjection;
 using Xunit;
