+++ tests/Abblix.Oidc.Server.UnitTests/Features/ExternalKeys/ExternalKeysWiringTests.cs
@@ -25,6 +25,7 @@ using System.Security.Cryptography;
 using Abblix.Jwt;
 using Abblix.Jwt.Signing;
 using Abblix.Oidc.Server.Common.Interfaces;
+using Abblix.Jwt.ExternalKeys;
 using Abblix.Oidc.Server.Features.ExternalKeys;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Options;
